Rollback legacy_deploy and Add tests for each revision(#12)
