+++ iconservice/deploy/icon_score_deploy_util.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 ICON Foundation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import os
+import time
+
+
+class DirectoryNameConverter:
+    counter = 0
+
+    @classmethod
+    def rename_directory(cls, path: str):
+        cls.counter += 1
+        os.rename(path, f"{path}{int(time.time()*10**6)}{cls.counter}_garbage_score")
